Implemented memory database alternative.

+++ Pzyber/Loki/Gomoku/DBType.java
@@ -12,13 +12,12 @@
  * http://www.gnu.org/licenses/gpl-3.0.html
  *
  * If you have questions, contact me at pzyber@pzyber.net
- *
  */
 
 package Pzyber.Loki.Gomoku;
 
 public class DBType {
-    public static final int MEMORY = 0;
-    public static final int FILE = 1;
-    public static final int SQL = 2;
+    public static final byte MEMORY = 0;
+    public static final byte FILE = 1;
+    public static final byte SQL = 2;
 }
